pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["AugustoCedric <cedric.pmi.rmp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
gitpython = "^3.1.43"
whitenoise = "^6.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


